msm: cpufreq: Add API to allow limiting of min and max cpu frequencies
CPUFreq provides scaling sysfs nodes to set min and max frequencies, but
there are times when the operating range has to be controlled
irrespective of the scaling limits, to ensure the thermal stability of
the device.

The API allows kernel thermal safeguards to pitch in and limit the
frequencies.

Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/cpufreq.c
Signed-off-by: flar2 <asegaert@gmail.com>

Conflicts:
	arch/arm/mach-msm/cpufreq.c
Change-Id: I3dc7496747792e6223695131f1644dd87054b025
Author:    Praveen Chidambaram <pchidamb@codeaurora.org>
Date:      Fri May 25 17:29:11 2012 -0600